black: collapse equivalent property-sets in output
Previously, property-sets resulting in the same mutant would be shown as
multiple lines in the output table.  This collapses them into a single entry.

This has the advantage of, when two property-sets have the same number of
surviving mutants, but actual different surviving mutants: they'll be shown on
the output as separate entries.

This also changes the order of the output:

was: nsurv, pset, mutant
is:  pset, nsurv, mutant